+++ src/main/kotlin/de/shyim/shopware6/inspection/twig/TwigBlockHashMissing.kt
@@ -5,10 +5,7 @@ import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiElementVisitor
-import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.util.indexing.FileBasedIndex
 import com.jetbrains.twig.elements.TwigBlockTag
-import de.shyim.shopware6.index.TwigBlockHashIndex
 import de.shyim.shopware6.inspection.quickfix.twig.AddMissingTwigVersioningCommentFix
 import de.shyim.shopware6.util.TwigUtil
 
@@ -25,18 +22,13 @@ class TwigBlockHashMissing : LocalInspectionTool() {
             return super.buildVisitor(holder, isOnTheFly)
         }
 
+        val filePath = virtualFile?.path ?: return super.buildVisitor(holder, isOnTheFly)
+        val chainPaths = TwigUtil.getExtendsChainPaths(holder.file)
+
         return object : PsiElementVisitor() {
             override fun visitElement(element: PsiElement) {
                 if (element is TwigBlockTag && element.name !== null && TwigUtil.getShopwareBlockComment(element) === null) {
-                    val filePath = element.containingFile.originalFile.virtualFile.path
-
-                    if (!FileBasedIndex.getInstance().getValues(
-                            TwigBlockHashIndex.key,
-                            element.name!!,
-                            GlobalSearchScope.allScope(element.project)
-                        )
-                            .any { it.relativePath == TwigUtil.getRelativePath(filePath) && it.absolutePath != filePath }
-                    ) {
+                    if (TwigUtil.getUpstreamBlocks(element.project, filePath, chainPaths, element.name!!).isEmpty()) {
                         return
                     }